use distinct types for master and outstation session errors
