+++ dnp3/src/entry/master/mod.rs
@@ -2,10 +2,10 @@ use crate::tokio::io::{AsyncRead, AsyncWrite};
 
 use crate::app::retry::ExponentialBackOff;
 use crate::master::handle::{Listener, MasterConfiguration, MasterHandle};
-use crate::master::session::MasterSession;
+use crate::master::session::{MasterSession, RunError};
 use crate::transport::TransportReader;
 use crate::transport::TransportWriter;
-use crate::util::task::{RunError, Shutdown};
+use crate::util::task::Shutdown;
 use std::future::Future;
 use std::io::Error;
 use std::time::Duration;
