+++ dnp3/src/util/task.rs
@@ -1,32 +1,6 @@
-use crate::link::error::LinkError;
-
 /// Indicates that a task shutdown has been requested
 #[derive(Copy, Clone, Debug)]
 pub struct Shutdown;
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum RunError {
-    Link(LinkError),
-    Shutdown,
-}
-
-impl From<LinkError> for RunError {
-    fn from(err: LinkError) -> Self {
-        RunError::Link(err)
-    }
-}
-
-impl From<Shutdown> for RunError {
-    fn from(_: Shutdown) -> Self {
-        RunError::Shutdown
-    }
-}
-
-impl From<crate::tokio::sync::oneshot::error::RecvError> for Shutdown {
-    fn from(_: crate::tokio::sync::oneshot::error::RecvError) -> Self {
-        Shutdown
-    }
-}
 
 pub(crate) struct Receiver<T> {
     inner: crate::tokio::sync::mpsc::Receiver<T>,
@@ -42,5 +16,11 @@ impl<T> Receiver<T> {
             Some(x) => Ok(x),
             None => Err(Shutdown),
         }
+    }
+}
+
+impl From<crate::tokio::sync::oneshot::error::RecvError> for Shutdown {
+    fn from(_: crate::tokio::sync::oneshot::error::RecvError) -> Self {
+        Shutdown
     }
 }
